requirements.txt
flask==2.1.0
flask-restx==0.5.1
flask-sqlalchemy==2.5.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
torch==1.10.2
torchvision==0.11.3
pillow==9.1.0
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
google-cloud-storage==2.3.0
werkzeug==2.1.1
flask-marshmallow==0.14.0
SQLAlchemy==1.4.36
flask-migrate==3.1.0
python-jose==3.3.0
bcrypt==3.2.0
email-validator==1.2.1
gunicorn==20.1.0